requirements.txt
paho-mqtt==1.3.1
mysql-connector==2.1.6
supervisor==3.3.3
django==1.11.9
django-bower==5.2.0
nodeenv==0.13.6
beautifulsoup4==4.6.0
